Re-do the fix for stacked graph highlighting
This reverts the change from https://github.com/danvk/dygraphs/pull/211
and instead provides a boolean arg to the plotter so that it can decide
not to draw if needed.

The problem with the approach from #211 is that this prevents error bars
from being highlighted, and those should be redrawn on mouseover.

Fixes issue 424.